% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bottom_legend}
\alias{bottom_legend}
\title{Bottom Legend Theme}
\usage{
bottom_legend()
}
\value{
ggplot2 theme
}
\description{
This theme sete legend in bottom and set map design.
}
\examples{
library(ggplot2)
map <- irProvinces()
ggplot(map)+
  geom_sf(aes(fill = area), size = 0.1, color = "white") +
  bottom_legend()
}
