% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{fullView}
\alias{fullView}
\title{Remove axis margin for Full View}
\usage{
fullView()
}
\value{

}
\description{
Add this as a layer to a ggplot object,in
order to remove axis margins.
}
\examples{
library(ggplot2)
map <- irProvinces()
ggplot(map)+
  geom_sf(aes(fill = area), size = 0.1, color = "white") +
  fullView()
}
