% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_raster_tile}
\alias{get_raster_tile}
\title{Get Map Tiles}
\usage{
get_raster_tile(map, provider, zoom)
}
\arguments{
\item{map}{an sf, sfc, bbox object.}

\item{provider}{the tile server from which to get the map. Providers contains:
"OpenStreetMap", "OpenStreetMap.MapnikBW", "OpenStreetMap.HOT", "Stamen.Toner",
"Stamen.TonerBackground", "Stamen.TonerLite", "Stamen.Watercolor", "Stamen.Terrain",
"Stamen.TerrainBackground", "Esri.WorldStreetMap", "Esri.DeLorme", "Esri.WorldTopoMap",
"Esri.WorldImagery", "Esri.WorldTerrain", "Esri.WorldShadedRelief", "Esri.OceanBasemap",
"Esri.NatGeoWorldMap", "Esri.WorldGrayCanvas", "CartoDB.Positron", "CartoDB.PositronNoLabels",
"CartoDB.DarkMatter", "CartoDB.Voyager", "CartoDB.VoyagerNoLabels", "OpenTopoMap", "HikeBike", "Wikimedia"}

\item{zoom}{the map zoom level between 1 to 20.}
}
\value{
data.frame with x,y and hex column
}
\description{
Get Map Tiles
}
\examples{
library(ggplot2)
raster_tile <- get_raster_tile(iran_border,provider = "Wikimedia",zoom = 7)
ggplot() +
geom_raster(data = raster_tile, aes(x, y, fill = hex)) +
geom_sf(data = iran_border,color = "blue",alpha = 0.1) +
scale_fill_identity() +
theme_void()
}
