% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irBorderProvinces}
\alias{irBorderProvinces}
\title{Iran's Border Provinces Map}
\usage{
irBorderProvinces()
}
\value{
sf object map
}
\description{
Iran's Border Provinces Map
}
\examples{
library(ggplot)
map <- irBorderProvinces()
ggplot(map) + geom_sf()
}
