% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irBorders}
\alias{irBorders}
\title{Iran's Border Map}
\usage{
irBorders()
}
\value{
sf object map
}
\description{
Iran's Border Map
}
\examples{
library(ggplot2)
map <- irBorders()
ggplot(map) + geom_sf() + theme_map()
}
