% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irCities}
\alias{irCities}
\title{Iran's Cities Map}
\usage{
irCities()
}
\value{
sf object map
}
\description{
Iran's Cities Map
}
\examples{
library(ggplot2)
map <- irCities()
ggplot(map) + geom_sf() + theme_map()
}
