% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irCounties}
\alias{irCounties}
\title{Iran's Counties Map}
\usage{
irCounties()
}
\value{
sf object map
}
\description{
Iran's Counties Map
}
\examples{
library(ggplot)
map <- irCounties()
ggplot(map) + geom_sf()
}
