% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irFeatures}
\alias{irFeatures}
\title{Iran's Natural Features Map}
\usage{
irFeatures()
}
\value{
sf object map
}
\description{
Iran's Natural Features Map
}
\examples{
library(ggplot2)
map <- irFeatures()
ggplot(map) + geom_sf() + theme_map()
}
