% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irProvinces}
\alias{irProvinces}
\title{Iran's Provinces Map}
\usage{
irProvinces()
}
\value{
sf object map
}
\description{
Iran's Provinces Map
}
\examples{
library(ggplot2)
map <- irProvinces()
ggplot(map) + geom_sf() + theme_map()
}
