% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{irRurals}
\alias{irRurals}
\title{Iran's Rurals Map}
\usage{
irRurals()
}
\value{
sf object map
}
\description{
Iran's Rurals Map
}
\examples{
library(ggplot2)
map <- irRurals()
ggplot(map) + geom_sf() + theme_map()
}
