% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{map_palettes}
\alias{map_palettes}
\title{Title}
\usage{
map_palettes(
  n,
  alpha = NULL,
  begin = 0,
  end = 1,
  direction = 1,
  palette = "BYR"
)
}
\arguments{
\item{n}{the number of generated colours for the palette.}

\item{alpha}{the opacity of the generated colours.}

\item{begin, end}{The interval within the palette to sample colours from.
Defaults to `0` and `1` respectively}

\item{direction}{Either `1` or `-1`. If `-1` the palette will be reversed}

\item{palette}{The name of the palette to sample from. See
[map_palette_names()] for a list of possible names.}
}
\value{

}
\description{
Title
}
\examples{
map_palettes(15, begin = 0.3, end = 0.6, palette = 'Earth')
}
\references{
[scico](https://github.com/thomasp85/scico)
}
