% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{map_palettes_show}
\alias{map_palettes_show}
\title{Show the different map palettes}
\usage{
map_palettes_show(palettes = map_palette_names())
}
\arguments{
\item{palettes}{One or more palette names to show}
}
\description{
This is a simple function to show a gradient of the different palettes
available in the `map` package
}
\examples{
map_palettes_show()
}
\references{
\href{https://github.com/thomasp85/scico}{scico}
}
