% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point2info.R
\name{point2Info}
\alias{point2Info}
\title{Add map meta data to points}
\usage{
point2Info(data, map, features = "province_code")
}
\arguments{
\item{data}{data.frame contains lng and lat columns}

\item{map}{sf polygons objects contain}

\item{feature}{string vector contains list of feature that must be added to data}
}
\value{
data.frame
}
\description{
Add map meta data to points
}
\examples{
point2Info(mountains,province,features= c("name"))
df <- data.frame(lng = c(51.42,59.55), lat = c(35.69,36.29))
point2Info(df, province, features = "name")
}
