% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map}
\alias{theme_map}
\title{Theme for Plotting Maps}
\usage{
theme_map(base_size = 12)
}
\value{
ggplot2 theme
}
\description{
This theme remove axis lines, major and minor line to see the maps better.
}
\examples{
library(ggplot2)
map <- irProvinces()
ggplot(map)+
  geom_sf(aes(fill = area), size = 0.1, color = "white") +
  theme_map()
}
