% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_elevation_tile}
\alias{get_elevation_tile}
\title{Get Elevation Map}
\usage{
get_elevation_tile(map, zoom)
}
\arguments{
\item{map}{Either a data.frame of x (long) and y (lat), an sp, sf, or raster object as input.}

\item{zoom}{The map zoom level between 1 to 20.}
}
\value{

}
\description{
Get Elevation Map
}
